Refactor in measurables page
